pyproject.toml
[project]
name = "food"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]
